Restructuring Conctrollers using mixings
